+++ Server/src/main/java/com/grade/forge/rubric/dto/RubricSubCriteriaRequest.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class RubricSubCriteriaRequest {
     private String description;
-    private Integer maxScore;
+    private Double maxScore;
     private Double weight;
 }
